Use the 'alternate' Grow Block City
The 'alternate' Grow Block City has matured enough that we can now
merge it into the main scene. More work to do still, but it's an
important milestone!

I did notice that the different part of the template buildings were
all visible by default. We want those offscreen buildings to start in
offscreen mode - and onscreen for the onscreen mode. So, I added a
function to check whether the node is currently offscreen or onscreen.

We now use that function when creating the buildings to set the
building's state appropriately. Part of me felt like we should have
done this as part of the template building's ready function, but I
prefer keeping that as sparse as possible. Besides, we can always
change it later.