Generate buildings using a consistent Random Number Generator
Previously, each factory stage was expected to resolve their own
random number needs - but I wanted it so that we could provide each
building with a seed, and get a consistent building out.

That meant that every stage would need a consistent way to generate
random numbers. After some thought, I decided it was best to provide
the stages with a premade random number generator. We would just seed
that, and then all the stages could use it at their leisure.