Add a light stage
Now we're building our different stages back up - the lights are a
critical part of our city, so it was important to me that we get it
in.

And actually, isolating it in a stage offered us some opportunity for
improvement. Since all the code was nice and isolated, it wasn't too
much overload to add a light-specific class: LightData! This allows
us to store all the light data in a single array, and store it
consistently.

Other than that, it was pretty much a straight copy.