Add a Roof Box stage
Next up was the the roof boxes - those little boxes that sit on top of
the buildings and help make for some more unique silhouettes. Not a
lot to say here; this step was somewhat similar in nature to the light
stage. We even made a custom roof data object to hand over data from
blueprint to construction.

Actually, one thing worth mentioning: it takes MUCH longer to make the
buildings now. It's about 6 seconds to blueprint, and maybe another
25 seconds for construction. That's pretty bad, considering we used to
be about 6 seconds to blueprint and about 1 second to construction.

If we do something similar to what we do with antennae, though, we
might be able to circumvent that...