Use a single mesh for all roof boxes
Previously, we were using an Auto Tower for the roof box. This wasn't
actually too bad when we were using the Footprint Building build
system, but the time overhead got a LOT worse when we switched to
using the Factory Stage build system.

However, we've already come up with a solution for this particular
problem: just have one mesh of a fixed size and we scale up the
MeshInstance node to whatever size we need. That's what we've been
doing with the Antennae, and it's worked pretty well.

In fact, I wasn't sure why we had used AutoTower originally - until I
started implementing this box thing. We had used Auto Tower to
generate meshes with no UV, which meant now windows. That broke the
second I started using the default mesh.

Forunately, we had the silhouette shader from the Antennae, which
operates like the Dot Window Light Shader BUT it has no window
texture! So I created three new materials to use this silhouette
shader, and everything works great!

Except I noticed some floating roof boxes. I thought it was a Y issue
until I experimentally lowered a box and saw that it was actually
jutting off the edge of the building! Some analysis showed that, where
I was supposed to be generating a Z offset, I was actually using the
formula for X. That explains the mishap, then. All fixed now, though!