Add an antennae stage
The last stage to add to the factory was the Antenna stage, where
we actually make the antennae! And, I have to say... I did not do a
good job setting up the antenna class/scene. It's frankly incomplete.
Seems like I meant to stick occurrence in their at some point but just
kinda... didn't? I'm gonna have to go back over that. At some point.

But, with all that settled, we're officially all in on the factory!
I deleted the Footprint Building - it was good while it lasted but, we
won't really need it. Now to see about fixing up that antenna scene...