Clean up the Antenna scene
As proud as I am of this project, I have started to tire of it in some
ways. Plus, work keeps me nice and exhausted most days. Wrap that
together with the complexity of the old Footprint Building and uh...
well, the Antenna was not exactly feature complete.

No occurrence, no enabling, no actual group functionality. I can't
believe I thought it was done! Honestly, after so many weeks of
working on it I guess I just kind of... FELT done.

This update improves things - we've now got the height ratio, the extra
scalar, the occurrence, and more. It's just not set up completely,
because it needs MCC integration. But, it's mostly complete! Now we
just need to do GUI stuff...

Also, I noticed we weren't clearing the groups in the Beacon scene - so
we do that now.