Update decorations more efficiently with "total" updates
Okay, so our whole MCC setup, while a bit complicated, has worked
pretty well. Everytime a key gets updated, we take certain actions.
Easy!

The one problem spot has been with decorations. See, most of our MCC
operations operate on resources, which then have follow-on visual
effects. Decorations, on the other hand, need to be manually updated
using group calls - in effect, a detatched way of calling a function.

The problem arose during a "mass key update" - which typically
happened when swapping profiles. We'd get a lot of often redundant
group calls - so many so that Godot was generating "too many calls"
errors! It was also causing ~20 seconds of hang time when switching
profiles, which was not acceptable to me.

To get around this, we now have a special variable to track whether
we're doing a mass update or not. If we're doing a mass update, we
just skip the group call functionality. In our mass key update
function, we then call the new "total update" function on each group
of decorations.

This new way of doing it is much more efficient - it means one
function call per decoration. In fact, if I was feeling lazy, I could
just compress all these separate keys into singular calls to the new
"total update" functions...