Add a control for the number of antennae per roof
I was trying to adjust the profiles when I realized that I couldn't
get the right number of antennae on the buildings. All I had was the
occurrence rating, which took away antennae at random. That worked
well BUT I was running into a situation where buildings had four or
five antennae, and the random nature of the occurrence rating wasn't
enough to get rid of these excess antennaes.

I realized I needed a separate control - a hard cap that allowed me to
set the maximum number of antennae on a building, so I could
sytlistically choose the maximum number of antennae associated with a
particular scheme.

Fortunately, with the Antennae scene, the Antennae factory stage, the
MCC, and all the MCC-GUI stuff - it was really easy! Seriously, it
only took like an hour! Now I'm thinking I can do something similar
for the rooftop boxes...