Add MCC Labels and Line Edits
Okay, I'm making one last feature push: we're gonna try saving files.
But, to save files, we need to be able to edit metadata and the like.
The author's name, the profile name, etc.

That means being able to show and display text. So for that, I've
created a MCC LineEdit class and MCC Label class. Technically the
label isn't really necessary, but I think it's cool, saves us having
to do an update manually.