Create the necessary user subdirectories
We're gonna need to load and save files to very particular user
subdirectories. To do that, we need to make sure those subdirectories
exist. I figure we can either do that at load-and-save, or we can do
that right when the program starts.

It made more sense to me to just create those directories right at the
beginning. Adds a certain amount of consistency, ya know? The one
downside is it's not runtime-resilient. So, if the user goes in there
and destroys the subdirectories, it can't recover until the program
restarts. That might be a bad idea... but I'm programming a bit too
late for good ideas, I think.

Anyway, the new UserDirEnforcer class is the one who handles this. I
just added it as an AutoLoad - easy! It's now the first thing that
happens in the program.