Add a scene for entering a file name and saving a file
Now that we've gotten the metadata sorted, we can actually save the
file. It's actually pretty straight forward - the user enters text
and we validate it. If the text proves to be valid, we enable the
save button. Once the user saves, we update the MCC and tell the
world that we saved the file.

My one complaint is that how Godot handles files is a little bit...
sloppy. Or maybe threadbare is a better word. As someone who works
mostly in some sort of off-color C, it kind of reminds me of the C
file API. You can open, you can close, but you need to check error
after all those things. I think I'd like to make a custom object to
wrap file, just to provide some conveince functions and improved error
checking.

Unfortunately that idea is running into the reality that I want to
tie a neat bow on this project (for now). Maybe we'll redo that file
stuff at some point. For now, this file will stay messy and C-like.