Load in user profiles
Saving user profiles - done! Loading user profiles? Hadn't done it
yet, so that was next up on the chopping block! And I have to say,
it was pretty easy!

That is, until I did the actual loading. Then the screen went black.
How could that be - how could that happen??? Well, that's the trouble:
it's the colors. They're being saved as some sort of list of float
values. I think, in all honesty, that they're just getting saved as a
raw string. That's pretty much useless to us. I need to think about
how to get around this one...