Actually use the directory search code I already wrote
I forgot I created a script to search directories for specific
extensions and return them as a list. My god, it was basically the
exact code segment I ended up copying. My foresight somehow outpaced
my hindsight - a rare occurrence, but not unheard of.

We now use the dedicated directory resource searcher in the Profile
Menu. Because using the strings as free-floating constants bugs me,
we now declare the paths in the Global Ref.

Also, I found an old Profile GUI widget we weren't using anymore, so I
deleted it!