Add a class for converting dicts to JSON and back again
Okay, so we we're having trouble because you can't save some Godot
base types in JSON. For example, colors don't get output as anything
resembling a color - just a string of floats. No good, no good!

So we need dedicated functions capable of encoding Color objects in a
JSON file; it also needed to be capable of de-encoding that
information from a JSON file and producing a color object.

In the face of that, I decided it'd be best to have our own dedicated
script to come with these functions. Enter, then, the Dictionary Load
and Save script. It loads and saves dictionaries! It has special
processing for saving and loading certain types. We'll expand these as
needed.

It all works almost seemlessly, except the windows. The building
windows don't update when you load a user profile. A cursory
investigation revealed that was because the building algorithm code is
being loaded as a float, not an int. Looks like we may need a way to
encode ints.