Add some load text so the user knows what's happening
I realized that, with nothing going on in those first three-or-four
seconds, the user probably isn't going to realize that something is,
indeed, happening. So I added some extra text using Godot's nifty
new 3D text. Unfortunately it seems to have an issue where it
disappears when it gets only halfway offscreen so I had to just get
rid of it once we were done loading.