Update the README
I hadn't updated the README. Like, ever. That was long overdue. So
I did what I called the basics - the introduction, the licensing, a
short (and kind of bad) overview of how the project works, as well
as a Special Thanks section.

I do feel bad about eulogizing Shamus in the Special Thanks section,
but genuinely this project would not exist without him. He was the
inspiration for all this. He's one of the key factors for me sitting
here writing this commit. Honestly, I figured it would be a glaring
ommission if I left it out.

However, there's no guarantee that everyone will see the README,
especially my credits section. I'll need to get that into the project
somehow.