Allow the user to resize the window
I don't have time right now to add proper options - settings to tune
the game up and down, or fixed resolutions, or even fullscreen
options. In the face of that, I have no choice but to allow the user
to freely resize the screen. They can really shoot themselves in the
foot that way but I don't think I have an option right now.