Update the README's licensing section
I realized that the MIT license, by it's very nature, does include
attribution in the form of my personal copyright. So I needed to
update the license section, where I said that attribution wasn't
required. That was just plain wrong.

So I changed it - just asking for a shoutout in the credits, or
something.