Use .import when running standalone
Okay, so previously, when iterating over directories, we ignored all
.import files, because we were looking for the files that were getting
imported (i.e. pngs).

Except, when you export the project and run standalone, all our pngs
disappear. If you list the directory, they're gone. BUT you can still
load the resource path!

So I had to add a workaround, because I'm kinda desperate to just
"crown" this program and call it done (for now). So, the MCC File GUI
widgets now offer an additional boolean: use import workaround. This
is passed directly to our Directory Resource Finder.

If that workaround is set to true, and we're running standalone, then
.import files now substitute for non .import files. Literally - we
just cut out the ".import" from the string. It's hacky and dirty, but
it works!