Create a workaround for light flickers on Windows
Okay, for some HORRID reason all of the lights start flickering on
Windows. I seriously can't figure out why this is happening. We're
getting no error messages at all! It has to be some sort of issue
underlying the Godot engine. Which is probably out of my scope to
fix.

Look, I'm not going to preach innocence here: I am spawning an
absolutely obnoxious number of omnilights. But only a given set are
ever visible onscreen at one time. That should undercut any issues,
and why am I not getting this on Linux? BAH!

My tester, Grant, showed me that switching off some lights fixed the
issue. If that's all it takes, FINE! I've literally added an if to
arbitrarily remove one of the lights from the Footprint Buildings,
IF we are in Windows. And I thought Godot was supposed to smooth
over these sorts of headaches!