Go back to the default window size
For whatever reason, my Windows is lower resolution than my Ubuntu
(must be upscaled or something). So when I went to start the project
on Windows, it frequently clipped off the screen and I had trouble
resizing it. It's easier to size up than size down, so I've reverted
to our original, smaller size.