Add & delete lights as they enter/exit
I'm trying to count down on the overall lights in the scene, and I had
this brilliant idea of deleting and adding omnilights. That would
minimize the total number of lights in the scene to only what we
needed.

It does have an effect - the Windows flashing issue is less frequent
and doesn't last as long. EXCEPT THAT IT'S STILL HAPPENING AAAAAGGGGH.

I think I'm going to keep this feature in because it's better practice
overall, but I need to bring this up with the Godot developers.
There's no getting around this.