Break down the project to a minimal reproduction state
In order to submit a bug report to the Godot Engine's GitHub, I need
to give them a "minimal reproduction project" that reproduces the
error with as little code as possible.

That's no easy task, considering the complexities involved in this
project. I had to make some new debug scenes - a debug building and
a debug city - and the MCC had to be gutted but otherwise kept in
place. Otherwise, it was mostly fat to trim.

I haven't tested this on Windows yet, so it's my genuine hope that
this demo runs without error. If it does run without error, then I
think there are some likely culprits: it's my shader (unlikely), it's
the threading (very unlikely) or it has to do with scaling the lights
and AABBs up-and-down (most likely, in my opinion). Or this is
actually just a Windows issue.