Show potential, spawned, and visible lights
Previously, we showed "overall" and "visible" lights. This was not
accurate. The "overall" lights actually counted the total number of
CityLight objects we had spawned, not how many Omnilights had been
spawned. Really it was a measure of how many lights COULD be spawned,
so I renamed it to "potential" lights.

Next, what was formerly "visible" lights actually measure how many
OmniLights had been spawned in, not how many were actually visible.
So I changed it to "spawned" lights.

That left us with a problem - I actually needed a way to measure the
visible lights. That was difficult for me to figure out - largely
because I refused to actually think about it and just kept guessing
and checking. Eventually I found the right sequence of incrementing
and decrementing to achieve the desired result. It's much rougher than
I like but this is a minimal reproduction project. It'll be fine.