Update max lights and scene local mesh
I got some feedback from a Godot maintainer - apparently, the "Max
Lights per Object" means "Max Lights per Resource". A revelation to
me, but it does actually make sense - so I've gone into the settings
and tuned it up to 109. I initially punched it up to 256, then to
512 - except 512 literally failed to compile for me. I'm afraid
others might not be able to compile the project as well, so I dialed
it back to 109; the tooltip says that's sort of an upper limit for
lowend devices.

The other recommendation was to make all of the CubeMesh instances
unique by making them "local to scene". It was so easy that I'd
barely believe it worked if it weren't for our total objects shooting
up to 10000 or so.

Finally, to reduce the amount of lighting on each DEMO building, I
changed the lights so they are basically placed at the corner of each
building. The height no longer has any input on the length of the
lights - it's a fixed constant now.

That didn't fix the issue, but we're incrementing ever closer!