Set up the AutoTower's UV values
Now that we're generating using the CubeMesh, we need to modify the UV
values - and possibly other values besides.

To do that, we needed to know which vertex was which. Some
experimentation showed that the vertices have this odd interleaved
scheme. So the vertices go FACE-A FACE-B FACE-A FACE-B and so on until
the eight total vertices for FACE-A and FACE-B are finished. Then we
do FACE-C FACE-D FACE-C FACE-D and so on and so on until we've got a
complete shape.

Unfortunately since we have face-specific processing, and this
alternating structure was so odd, I had no choice but to just list out
all the indicies and handle them appropriately. Maybe there's a
mathematical way of grouping it but I certainly couldn't think of it.

Anyway, the structure of the UV values in the CubeMesh was odd as
well - I think they were deliberately structured to sort of wrap
around the Cube. This meant we had to handle each face differently to
get the exact UV values we needed. Kinda ugly code, but it works.

Now I need to start working with the UV2 values to pack in the light
data... or light reference data. Or whatever.