Pack light texture coordinates into UV2
We got the UV stuff all set up, now we need to set up the UV2 stuff.
The overriding theory at play here is that we'll have some sort of
noise texture, and we'll pick values from that using the UV2
coordinates.

That's the theory anyway - I'm kinda making this up as I go. Best that
I can do, I suppose.

Anyway, we now generate UV2 values for each vertex based on the
vertex's position in window units. To do this, I added new LIGHT NOISE
UV constants. Everything seems to be working so far...