Use a power-based algorithm and color adding for lights
Okay so our previous algorithmic outing used this kind of loony
distance-based algorithm. Like, I was using the noise texture for
the distance. Surprisingly, IT DIDN'T WORK! Or it did work, but it
looked weird.

The issue was two fold - first off, turns out you're supposed to ADD
light's colors in, not mix them. My bad. I fixed that.

Secondly, I decided to actually USE the power-based algorithm. You
multiply the color by the power, which you multiply by the
attenuation.

And actually, on that point, I changed the distance calculation to
actually use distance. To scale that distance, I added a distance
coefficient. We'll have to use this field for scaling because window
units are so much smaller than actual units.

It looks okay if the colors are largely the same, but if the colors
are different it turns into this sort of unpalatable soup. I actually
got rid of the contrast controls, but I think it's time to bring them
back...