Separate lights by adjusting the noise contrast
There was too much light blobbing going on with the previous shader
iteration - so I've adjusted the noise in the texture. This forces the
colors towards white-or-black, changing the noise texture into a
series of "color islands". The effect looks pretty good. Not as good
as using omnilights, but we know why we can't do that.