Mix the lights with the windows according to our factors
There was one final bit of functionality that was missing from the
power shader: mixing the light colors with the window dots.

Fortunately, that was actually pretty easy. Just determine the
appropriate mix factor and multiply it by the color at add-time.

Also, I took the opportunity to add a bunch of comments in the shader.
There was a bunch of stuff left unsaid and I'm sure someone
particularly clever could have figured it out, but I'm not
particularly clever so I did not take that risk.