Integrate the power window shader into the city program
Alright, enough messing around with theory - time to put this power
shader into the program and see how it looks!

... Not great. Really kinda bad actually. But not hopelessly bad! I
mean some of it is just bad because only the buildings are different
but the silhouette stuff isn't. It's the mixing of the lights and how
they look, and how they just seem to extend forever.

I can change that though! I just need to tinker with the shader
uniforms, the shader code, and the auto building. The shader uniforms
are the control levers, so that should be where I start. But, I'll
need to update the GUI to make that happen.