Play around with noise textures
Okay, I wanted to mess around with the noise textures we were using.

I haven't really been enjoying the textures we've been using; the
noise stuff just doesn't look that great. I wanted to have some more
fine control over the textures and, to be honest, the generated
textures were adding some significant load time.

So I changed the shader around - each light now has a shift and a
texture scale. The idea is we can use the same texture for each light
and then use scaling and shifting to move it around appropriately. To
that end, I added some noise textures to help explore what different
noise textures could produce. I'm particularly fond of the cell noise
texture - that one has really good results.

Anyway, I also added a "scalar" and "coefficient" for each light. The
idea is that the scalar is the user's handle on each light, to scale
up the power - the coefficient is for turning the colors on and off
and should really only be set to 0 or 1.

Since I had so many light variables now, I shortened a bunch of
variables from "light" to lx - i.e. LightOne to L1.