Experiment with a range-based shader (it doesn't work)
Okay so the whole power shader thing wasn't working out, so I found
some code online where somebody was calculating attenuation using the
radius of a light and the distance from it. "Genius!", I thought to
myself. "That's what I'll use the noise texture for!".

So I went through all the work crunching the numbers to make something
that used this radius-range technique. And it LOOKS EXACTLY THE SAME
ARRRRRGGGGHHHHHHHHHHHH

Okay, it's time to go about this a different way. Each building is
going to need four lights, and four lights only. The question is, how
do I pack that data in to the available data pools - and how do I make
that consistent when creating decorations and the like?