Use normals in the Packed Light Shader
The lights were mixing too much, and it just didn't look good.
Everything was turning into color mush. The brilliance of the original
city lighting algorithm was the way the colors just occassionally
shone across the buildings - a sea of different colors instead of just
the same one over and over.

So, after a lot of strain and pain, I figured out how to actually
implement the lighting in the fragment shader. That's a stupid
sentence and it makes me angry but sometimes that's just how it's got
to be. And as painful as it was, I have to say, I'm learning a lot!
Feel like I've really got a handle on this shader nonsense.

One thing worth mentioning here: I couldn't get the NORMAL built-in of
the fragment shader to work. I really don't know why. The lights were
appearing and disappearing depending on the view angle of the camera.
I assumed this meant it was in viewspace, but my technique for
converting out of viewspace didn't seem to actually work. I may have
made a typo though - like I said, a lot of this was pretty painful. So
I resorted to just taking the NORMAL out of the VERTEX shader. Lazy,
inefficient, and functional. The best sort of code!