Pack light information into the vertex color array
The idea for the Packed Light shader was (perhaps unsurprisingly) that
we would pack information about the light into some part of the array
mesh. I figured we each light would need a group, a range, and a xz
position. So that was an int, a float, and a Vector2.

Except all I had was the TANGENTs (which is 3 floats), the COLORs
(which is 4 floats between 0 & 1), and the UV2 (which is 2 floats).

Floats aren't good for packing values, so NONE of these are good
choices. But they're our only choices. We don't have an int or
something we can use instead.

Given that, I felt that the Vertex Color array was the best candidate.
The only problem was that it had to contain values between 0 and 1.
We'd have to divide our int bit mask by a certain value to get it
between 0 and 1. Godot's documentation stated the floats were
generally accurate up to six figures.

Our bit-packed value must then be at most 999,999. We would divide by
one million to turn it into a float between 0 and 1. Consulting a
table of the powers of 2, the highest power of 2 that still fit under
999,999 was 524,288. That's 2^19, which means we had 19 bits to choose
from.

First we had to assign the group. We have four groups, so that's two
bits. Easy.

Next I allocated the range. Our buildings are generally between 20 and
118 blocks tall, so I decided we could have a max value of 127 - 2^7.
So 7 bits for building height.

9 bits used, 10 bits left. Except I still had to do x and z positions.
Each one of those would need to be 5 bits then. That meant the max
value was 31. That didn't feel right, though I figured I could
multiply it by a value in the shader to max the effective max value
96.

I decided that I should implement this to test it out - and it doesn't
work. Not only does it not work, but our values are coming out
truncated due to floating point inaccuracy. Jeez.

It looks like 6 digits is inaccurate, so we should use 5 digits
instead. The highest power of 2 thats 5 digits is 65,536: 2^16. So
now I'll only have 16 bits to use! Maybe I can offload some stuff into
another array, like UV2...