Only pass group and range into the packed light shader
For all attempts to try and make the position and group fit into a
single, it just didn't work. I wish I knew why, but I don't think
there's a good way of encapsulating that. The truth is, far and away
as I can tell, Godot just mangles floats via compression.

That's not bad if you're doing something normal. I was not doing
something normal. It was bit-masking floats! OF COURSE IT WASN'T
GOING TO WORK!

Except, it only really seemed to be mangling the color values. The UV2
values the ranges seemed to be working fine? Bah!

So I decided to use the color value JUST for the group. I reverted to
the wacky "positive corner" idea that I had earlier - it was wacky and
didn't make a lot of sense but it WORKED which is more than I could
say for anything else. The color is now a multiplied by 4 to get the
group.

But the UV2 is starting to give me issues. The northern ranges are
clipping the southern ones causing weird lights to appear and
disappear at given ranges!

Say, I just had an idea - I've been trying to bitpack information. But
the floats are stored by value... what if instead of bit packing, I
did value packing? The 1's would be the group, and then the 10's,
100's, and 1000's column could be the range? Then everything would fit
in the Vertex Color! That just might work...