Pack range and group by value, not by bit
Last commit, I speculated that since floats are compressed/stored by
value, if I used "value packing" (i.e. each digit standing for a
different value) then I could pass values much more consistently than
when I was trying to pack information into each bit.

According to this theory, I could really use any values I fancied. I
knew there were only four groups, so we'd only need one digit to
encode one light's group. The question of range was trickier - I knew
from previous commits that 5 significant digits was about all we could
expect to be accurate. So, to not push it, I decided to cap the range
at 999. That meant all the information we needed could be contained in
four digits!

I experimented with sticking range and group all in the vertex COLOR
array (that would be cleaner) but I found that was generating weird
group inaccuracies. That accursed float inaccuracy in the least
significant digit was striking once again!

I moved the group to the UV2 since I figured I could get away with
storing a 2 digit number in there without inaccuracy. Except I was
still getting inaccuracy in the least significant digit. After some
analysis, I saw that it was sort of changing the value in an attempt
to compensate for a POSSIBLE floating point error.

As an example, let's say the value-packed float was 0.24. I multiplied
this by 100 and was getting 23.99999. See? That 4 was getting rounded
into a 3.

My solution, dumb as it was, was to add a .00001 to each group float.
So 0.24 becomes 0.2402 (the add happens twice, it doesn't really change
anything). Multiply by 100 and we get something like 24.029999! Digit
preserved!

So FINALLY, after MUCH STRUGGLE, I am passing light information into
the building and the shader is actually reproducing it!

There is one odd thing I noticed: the range only really seems to
happen by 4s. For example, the range values [32, 33, 34, 35] all seem
to just be... 32? When you hit 36 it suddenly snaps up to 36 but then
doesn't move for the next 4 values. I'm thinking this is more floating
point inaccuracy but at this point... who cares? It's good enough!

I briefly thought about attempting to pack everything into the VERTEX
COLOR array but given the inaccuracy issues with just the range, I
think this solution is good enough!