Determine the light information in the factory
Okay, now that we've got the lights working, we need to actually
generate values and then pass them into the AutoTower.

Fortunately, we still had the Light Stage knocking around. Since these
new lights are acutally simpler than the actual lights, it was really
easy to get everything set up.

The only thing that was odd was that the "range" values aren't exactly
what I expected them to be, so I increased the size of possible values
to almost double the possible building size. It seems the color is
actually having a big impact... Maybe I should have additional sliders
for power and range? I think that could work.