Use Godot 4.0 Beta 6's attenuation algorithm
I wasn't really happy with how the attenuation algorithm looked. What
I really wanted was for the lights to mirror how they used to look
when we were using Godot lights.

It ocurred to me - what was stopping me from just copying Godot's
algorithm? I guess I could've done that all along, I just never
bothered looking into it! So I pulled the master source, grep'd it,
and found that omni attenuation was in a function - I just copied
that into my code, with some slight modification.

Although the lights are not as prominent as I'd like on taller
buildings (it seems the lights don't extend as far as I'd like), it
looks MUCH better on the smaller buildings. It adds slight color
variation to the taller buildings and overall gives the scene MUCH
more depth and texture. It's just what it needed!

Now I can go add these lights to the decorations...