Add a silhouette version of the packed light shader
I was going to work on some light-compliant decorations: the antennae
and the roof boxes.

The silhouette shader itself was easy - it's basically just the
regular packed light shader but with all the Dot-Window code scrubbed
out.

I then set about updating the actual decorations code when I ran into
a problem - it just ran like crap! See, my GDScript mesh building
isn't the fastest thing in the world. That's acceptable for the
buildings, but for ALL those antennae and rooftop boxes? Forget about
it! It was taking upwards of two minutes to do the initial generation,
and then up to 30 seconds when transitioning between profiles. That's
useless to us.

I don't know what to do about it, though. I guess we'll just have to
be sparing with how we use the decorations. Since the decorations
don't receive any proper light, they're just defaulting to the darkest
possible mode. That creates a sharp and ugly contrast ESPECIALLY in
brighter schemes. I guess there's nothing to do but try and work
around it!