Add a new texture-data dependent light shader
After Thanksgiving and Christmas, a vacation, and the Super Bowl, I am
finally back to the neon-city project. And just in time for the
February deadline I set for myself. Or just in time to PANIC about the
deadline I set for myself.

Anyway, so the trouble with the previous shader was that packing all
that information in was far too information rich. Slowed the program
down. ESPECIALLY for all of our decorations. And we can't be without
our decorations - they really tie the whole city together!

The solution I came up with was to basically move the packing -
instead of packing everything into a building, we pack it into a
texture and just look at it. Not exactly innovative but GOOD ENOUGH!
We make use of all 3 channels - RGB - and generate this in GIMP via
the "Hurl" filter (which is why the next texture is "hurl noise").

The new shader has variation for both range and offset. It's
unforunate that we can't get the lights to scale proportionally to the
buildings but frankly nothing is perfect here.

The problem I'm having is making everything look good - it's just a
few too many levers. Or rather, the offset and range variation need to
be set PER COLOR SCHEME. Some schemes need less range, or more offset;
some need more variance and others less. Think I'm gonna have to cook
up a GUI...