Use the building height to determine light range
Ever since I've started working on alternative lighting methods, the
buildings have never looked as they good with the brute-force
omnilights. I'm sure there's lots of reasons for this, but one of the
few I've been able to articulate is the range of the lights.

See, range used to be set per-building: the omnilights were always
paired to a building and then configured around that building's
height. Within a random range, of course. This ensured a good-looking
color-fade on the building regardless of height. I figured getting a
feature like that into the latest shader would be worth something.

Of course, that meant I had to figure out how to pack a building's
height into the building mesh. This is something I wanted to avoid:
information packing is what doomed the last set of shaders when I
couldn't sufficiently handle all the decorations in a timely manner.

So how could I pass in per model information without editing the
vertices? After some thought, I realized that the model's origin in
realspace was the same for all vertices. Ergo, we could use the Y
offset of that origin as the building's height!

My dumb idea had one catch though: since the origin had to be at the
building's height, that meant we had to create the model such that the
center of the roof was at origin and the building projected DOWNWARD.
That required a whole lot of changes! The AutoTower now constructs the
building mesh in such a way and shifts it appropriately so it APPEARS
to start at origin. Genius! Or stupid. Stupid genius? Yeah, something
like that.

It's not without it's issues though. I wanted to capture what I had
done before tackling these new problems.

First off, the lights show up in the editor in the AutoTower scene,
but NOT in the actual program at runtime. Not sure why that is.

Secondly, there's seems to be an issue with mixing the colors. I'm
noticing that the lighting colors are overwriting each other instead
of properly blending. Maybe my discount bumpkin math has finally
caught up with me?