Dynamically handle scale in the light shader
Whenever we ran the main city scene, the buildings displayed no color.
After some experimentation, I figured out that it was due to the
scaling we were applying to the buildings. I can't place exactly what
was going wrong, but it definitely had to do with the origin being the
same in MODEL space but being scaled up in GLOBAL space.

My solution, then, was to "unscale" the building height. After some
thought-work, I realized we could calculate the scale by comparing the
VERTEX in MODEL space and GLOBAL space. If they were both offsets from
an origin, the the relationship between them SHOULD be the scale.

Took a lot of playing around, but I did get it working - and I'm
pretty happy with the results! This is the best the city has looked in
months. Now to see if I can get the colors to blend better...