Lower the decay factor to soften the lighting
I thought there was something wrong with how the light colors were
blending. On the test material, the pure yellow seemed to be
completely overwhelming all of the other colors.

Turns out that's because pure yellow is full R & G. So it completely
absorbs any R & G, meaning the light color really only changes
whenever Blue is added to the equation. And I wasn't doing that with
my test materials.

Still, we want to have more mixture of colors than we were seeing. And
the lights were too bright! So I toned the decay factor until the
lights were softer, encouraging more mixing.